chore: yorkes+pbio pr review changes
