+++ typescript/cli/src/check/warp.ts
@@ -1,79 +1,41 @@
-import { warpConfigToWarpAddresses } from '@hyperlane-xyz/registry';
-import {
-  EvmERC20WarpRouteReader,
-  WarpCoreConfig,
-  WarpRouteDeployConfig,
-  attachContractsMapAndGetForeignDeployments,
-  hypERC20factories,
-  proxiedFactories,
-} from '@hyperlane-xyz/sdk';
-import { objMap, promiseObjAll } from '@hyperlane-xyz/utils';
+import { stringify as yamlStringify } from 'yaml';
 
-import { CommandContext } from '../context/types.js';
-import { logRed } from '../logger.js';
-import { ViolationDiff, diffObjMerge } from '../utils/output.js';
+import { WarpRouteDeployConfig, normalizeConfig } from '@hyperlane-xyz/sdk';
+import { ObjectDiff, diffObjMerge } from '@hyperlane-xyz/utils';
+
+import { log, logGreen } from '../logger.js';
+import '../utils/output.js';
+import { formatYamlViolationsOutput } from '../utils/output.js';
 
 export async function runWarpRouteCheck({
-  context,
   warpRouteConfig,
-  warpCoreConfig,
+  onChainWarpConfig,
 }: {
-  context: CommandContext;
   warpRouteConfig: WarpRouteDeployConfig;
-  warpCoreConfig: WarpCoreConfig;
-}): Promise<[ViolationDiff, boolean]> {
-  const warpAddressesByChain = warpConfigToWarpAddresses(warpCoreConfig);
-  const { foreignDeployments } = attachContractsMapAndGetForeignDeployments(
-    warpAddressesByChain,
-    { ...hypERC20factories, ...proxiedFactories },
-    context.multiProvider,
-  );
-
-  // Check if there any non-EVM chains in the config and exit
-  const nonEvmChains = Object.keys(warpAddressesByChain).filter(
-    (c) => foreignDeployments[c],
-  );
-
-  if (nonEvmChains.length > 0) {
-    const chainList = nonEvmChains.join(', ');
-    logRed(
-      `${chainList} ${
-        nonEvmChains.length > 1 ? 'are' : 'is'
-      } non-EVM and not compatible with warp checker tooling`,
-    );
-    process.exit(1);
-  }
-
-  const warpCoreConfigByChain = Object.fromEntries(
-    warpCoreConfig.tokens.map((token) => [token.chainName, token]),
-  );
-
-  const onChainWarpConfig = await promiseObjAll(
-    objMap(warpRouteConfig, async (chain) => {
-      return new EvmERC20WarpRouteReader(
-        context.multiProvider,
-        chain,
-      ).deriveWarpRouteConfig(warpCoreConfigByChain[chain].addressOrDenom!);
-    }),
-  );
-
+  onChainWarpConfig: WarpRouteDeployConfig;
+}): Promise<void> {
   // Go through each chain and only add to the output the chains that have mismatches
-  const violationsByChain = Object.keys(warpRouteConfig).reduce(
+  const [violations, isInvalid] = Object.keys(warpRouteConfig).reduce(
     (acc, chain) => {
-      const [merged, isInvalid] = diffObjMerge(
-        onChainWarpConfig[chain],
-        warpRouteConfig[chain],
+      const { mergedObject, isInvalid } = diffObjMerge(
+        normalizeConfig(onChainWarpConfig[chain]),
+        normalizeConfig(warpRouteConfig[chain]),
       );
 
       if (isInvalid) {
-        acc[0][chain] = merged;
+        acc[0][chain] = mergedObject;
         acc[1] ||= isInvalid;
       }
 
       return acc;
     },
-    [{}, false] as [{ [index: string]: ViolationDiff }, boolean],
+    [{}, false] as [{ [index: string]: ObjectDiff }, boolean],
   );
 
-  return violationsByChain;
+  if (isInvalid) {
+    log(formatYamlViolationsOutput(yamlStringify(violations, null, 2)));
+    process.exit(1);
+  }
+
+  logGreen(`No violations found`);
 }
